requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
click==8.1.3
crispy-bootstrap4==2022.1
cssselect2==0.7.0
dj-database-url==1.3.0
Django==4.1.6
django-appconf==1.0.5
django-bootstrap5==23.1
django-crispy-forms==2.0
django-heroku==0.3.1
django-mathfilters==1.0.0
django-safedelete==1.3.2
django-select2==8.1.1
docopt==0.6.2
frozenlist==1.3.3
gunicorn==20.1.0
html5lib==1.1
lxml==4.9.2
multidict==6.0.4
openai==0.27.2
oscrypto==1.3.0
Pillow==9.4.0
pipreqs==0.4.11
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pypdf==3.7.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
qrcode==7.4.2
reportlab==3.6.12
sqlparse==0.4.3
svglib==1.5.1
tqdm==4.65.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
whitenoise==6.4.0
xhtml2pdf==0.2.9
yarg==0.1.9
yarl==1.8.2
